Don't keep the Java VM loaded if not needed
The problem is that the JVM cannot be easily unloaded even
when it's used only to get the available JDBC drivers list.

The current solution is to fork a child, load the JVM from
the child, get the available JDBC drivers list, transmit that list
to the parent process and quit the child.

If a connection to a JDBC is actually required, then the JVM will
be re-loaded from the main process.