sql-parser: Enum Macros and get_type() functions: Fix the prefix.
libgda/sql-parser/Makefile.am: glib-mkenums was called with a
GDA_SQL_PARSER_TYPE prefix instead of a GDA_TYPE prefix, though
the enum types themselves have regular Gda prefixes, even if they
happen to be in a sub-directory. This made it hard for
language-bindings (including libgdamm and pygda) to guess the
function and macro names.

For bug #617429