GdaTime: ported to GDateTime; default to UTC
GdaTime now internally holds a GDateTime, set
to epoch date 1970-1-1. When saving data to
data base it defaults to UTC time, because that
is the default on SQLite and may other applications.

Added new API to GdaTime, so for UI is possible
to show time always in local time, UTC or with time
zone. Is possible to manipulate time using GTimeZone.