mdb-provider: Pass MdbHandle to the mdb_set_date_fmt() call
mdbtools >= 0.9.0 require a MdbHandle to be passed to the
mdb_set_date_fmt() call. Move the call to directly after mdb_open()
(when we first get the handle) and pass the handle to it.

Note this change means that the mdb-provider will now no longer
compile with older mdbtools versions. To enforce this the
pkg-config check now checks for mdbtools >= 0.9 and the fallback
to manual checking in /*/lib is removed, except for the case where
the user has specified a path to check on the cmdline.